requirements.txt
asgiref==3.5.2
Django==3.2.15
pytz==2022.2.1
sqlparse==0.4.2
typing_extensions==4.3.0
django-crispy-forms==1.14.0
Pillow==9.2.0
joblib==1.1.0
roboflow==0.2.14
